+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## [1.0.0] | 17.3.2021
+### Added
+- Added settings to manage guild
+    - Prefix: To change guild prefix
+    - AutoClean: To choose whether to delete the command message
+  
+- Added Info command to give info about the current track
+
+### Fixed
+- Lyrics command now handles the char cap that discord has
+
+### Gradle Import
+- Lavaplayer Unofficial from 1.3.75 to 1.3.76
+  - Fixes the spotify authentication bug
+
+
 ## [0.1.0 Alpha] | 17.3.2021
 ### Initial release
 - It should be pretty stable but not perfect.